First part of 64bit support
Stepped cmake up to version 3.1 (removed local version check, only needed in master cmake file)
With version 3.1 is is now possible to do (in build)
cmake -G "Visual Studio 12" -A x64 ..
if the -A parameter is omitted, 32bit is assumed.
All CMake files and a couple of compiler warnings has been changed.
It does NOT link in 64bit, there is a minizip library (for second part).

+++ DocFormats/core/src/xml/DFXML.c
@@ -139,7 +139,7 @@ static void SAXStartElementNS(void *ctx, const xmlChar *localname,
         const xmlChar *attrURI = attributes[i*5+2];
         const xmlChar *attrValueStart = attributes[i*5+3];
         const xmlChar *attrValueEnd = attributes[i*5+4];
-        unsigned long attrValueLen = attrValueEnd - attrValueStart;
+        unsigned long attrValueLen = (unsigned long)(attrValueEnd - attrValueStart);
 
         Tag attrTag = DFNameMapTagForName(parser->document->map,(const char *)attrURI,(const char *)attrLocalName);
         const TagDecl *attrTagDecl = DFNameMapNameForTag(parser->document->map,attrTag);
